+++ internal/credentials/access_error.go
@@ -94,10 +94,12 @@ func AccessError(msg string, err error, opts ...authErrorOption) error {
 	buffer.WriteString(msg)
 	buffer.WriteString(" (")
 	for i, opt := range opts {
-		if i != 0 {
-			buffer.WriteString(",")
+		if opt != nil {
+			if i != 0 {
+				buffer.WriteString(",")
+			}
+			opt.applyAuthErrorOption(buffer)
 		}
-		opt.applyAuthErrorOption(buffer)
 	}
 	buffer.WriteString("): %w")