+++ internal/query/session.go
@@ -9,6 +9,7 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Query"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/allocator"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/options"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/query"
 )
@@ -19,7 +20,7 @@ type Session struct {
 	id          string
 	nodeID      int64
 	queryClient Ydb_Query_V1.QueryServiceClient
-	status      query.SessionStatus
+	status      sessionStatus
 	close       func()
 }
 
@@ -70,14 +71,18 @@ func (s *Session) NodeID() int64 {
 	return s.nodeID
 }
 
-func (s *Session) Status() query.SessionStatus {
-	return query.SessionStatus(atomic.LoadUint32((*uint32)(&s.status)))
+func (s *Session) sessionStatus() sessionStatus {
+	return sessionStatus(atomic.LoadUint32((*uint32)(&s.status)))
+}
+
+func (s *Session) Status() string {
+	return s.sessionStatus().String()
 }
 
 func (s *Session) Execute(
-	ctx context.Context, q string, opts ...query.ExecuteOption,
+	ctx context.Context, q string, opts ...options.ExecuteOption,
 ) (query.Transaction, query.Result, error) {
-	tx, r, err := execute(ctx, s, s.queryClient, q, query.ExecuteSettings(opts...))
+	tx, r, err := execute(ctx, s, s.queryClient, q, options.ExecuteSettings(opts...))
 	if err != nil {
 		return nil, nil, xerrors.WithStackTrace(err)
 	}