+++ internal/xsql/dsn_test.go
@@ -13,8 +13,10 @@ func TestParse(t *testing.T) {
 	newConnector := func(opts ...ConnectorOption) *Connector {
 		c := &Connector{}
 		for _, opt := range opts {
-			if err := opt.Apply(c); err != nil {
-				t.Error(err)
+			if opt != nil {
+				if err := opt.Apply(c); err != nil {
+					t.Error(err)
+				}
 			}
 		}