+++ options.go
@@ -247,9 +247,9 @@ func With(options ...config.Option) Option {
 // MergeOptions concatentaes provided options to one cumulative value.
 func MergeOptions(opts ...Option) Option {
 	return func(ctx context.Context, c *Driver) error {
-		for _, o := range opts {
-			if o != nil {
-				if err := o(ctx, c); err != nil {
+		for _, opt := range opts {
+			if opt != nil {
+				if err := opt(ctx, c); err != nil {
 					return xerrors.WithStackTrace(err)
 				}
 			}
@@ -465,6 +465,25 @@ func WithTraceTable(t trace.Table, opts ...trace.TableComposeOption) Option { //
 				append(
 					[]trace.TableComposeOption{
 						trace.WithTablePanicCallback(c.panicCallback),
+					},
+					opts...,
+				)...,
+			),
+		)
+
+		return nil
+	}
+}
+
+// WithTraceQuery appends trace.Query into query traces
+func WithTraceQuery(t trace.Query, opts ...trace.QueryComposeOption) Option {
+	return func(ctx context.Context, c *Driver) error {
+		c.queryOptions = append(
+			c.queryOptions,
+			queryConfig.WithTrace(&t,
+				append(
+					[]trace.QueryComposeOption{
+						trace.WithQueryPanicCallback(c.panicCallback),
 					},
 					opts...,
 				)...,