+++ query/client.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/closer"
-	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/options"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -16,7 +16,7 @@ type Client interface {
 	// - retry operation returned nil as error
 	//
 	// Warning: if context without deadline or cancellation func than Do can run indefinitely.
-	Do(ctx context.Context, op Operation, opts ...DoOption) error
+	Do(ctx context.Context, op Operation, opts ...options.DoOption) error
 
 	// DoTx provide the best effort for execute transaction.
 	//
@@ -30,7 +30,7 @@ type Client interface {
 	// If op TxOperation returns nil - transaction will be committed
 	// If op TxOperation return non nil - transaction will be rollback
 	// Warning: if context without deadline or cancellation func than DoTx can run indefinitely
-	DoTx(ctx context.Context, op TxOperation, opts ...DoTxOption) error
+	DoTx(ctx context.Context, op TxOperation, opts ...options.DoTxOption) error
 }
 
 type (
@@ -49,25 +49,20 @@ type (
 
 		Session
 	}
-
-	DoOption interface {
-		applyDoOption(o *DoOptions)
-	}
-
-	DoOptions struct {
-		Label        string
-		Idempotent   bool
-		RetryOptions []retry.Option
-		Trace        *trace.Query
-	}
-
-	DoTxOption interface {
-		applyDoTxOption(o *DoTxOptions)
-	}
-
-	DoTxOptions struct {
-		DoOptions
-
-		TxSettings TransactionSettings
+	bothDoAndDoTxOption interface {
+		options.DoOption
+		options.DoTxOption
 	}
 )
+
+func WithIdempotent() bothDoAndDoTxOption {
+	return options.WithIdempotent()
+}
+
+func WithTrace(t *trace.Query) bothDoAndDoTxOption {
+	return options.WithTrace(t)
+}
+
+func WithLabel(lbl string) bothDoAndDoTxOption {
+	return options.WithLabel(lbl)
+}