+++ trace/details.go
@@ -44,6 +44,11 @@ const (
 	TablePoolSessionLifeCycleEvents
 	TablePoolAPIEvents
 
+	QuerySessionEvents
+	QueryExecuteEvents
+	QueryTransactionEvents
+	QueryPoolEvents
+
 	TopicControlPlaneEvents
 
 	TopicReaderCustomerEvents
@@ -89,6 +94,11 @@ const (
 		TablePoolLifeCycleEvents |
 		TablePoolSessionLifeCycleEvents |
 		TablePoolAPIEvents
+
+	QueryEvents = QuerySessionEvents |
+		QueryPoolEvents |
+		QueryExecuteEvents |
+		QueryTransactionEvents
 
 	TablePoolEvents = TablePoolLifeCycleEvents |
 		TablePoolSessionLifeCycleEvents |
@@ -144,6 +154,12 @@ var (
 		TablePoolSessionLifeCycleEvents: "ydb.table.pool.session",
 		TablePoolAPIEvents:              "ydb.table.pool.api",
 
+		QueryEvents:            "ydb.query",
+		QueryPoolEvents:        "ydb.query.pool",
+		QuerySessionEvents:     "ydb.query.session",
+		QueryExecuteEvents:     "ydb.query.execute",
+		QueryTransactionEvents: "ydb.query.tx",
+
 		DatabaseSQLEvents:          "ydb.database.sql",
 		DatabaseSQLConnectorEvents: "ydb.database.sql.connector",
 		DatabaseSQLConnEvents:      "ydb.database.sql.conn",
@@ -191,9 +207,9 @@ func MatchDetails(pattern string, opts ...matchDetailsOption) (d Details) {
 		err error
 	)
 
-	for _, o := range opts {
-		if o != nil {
-			o(h)
+	for _, opt := range opts {
+		if opt != nil {
+			opt(h)
 		}
 	}
 	if h.posixMatch {